+++ noodles-bam/examples/bam_count_async.rs
@@ -13,8 +13,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let src = env::args().nth(1).expect("missing src");
 
     let mut reader = File::open(src).await.map(bam::AsyncReader::new)?;
-    let header = reader.read_header().await?.parse()?;
-    reader.read_reference_sequences().await?;
+    let header = reader.read_header().await?;
 
     let mut records = reader.record_bufs(&header);
     let mut n = 0;