+++ noodles-bam/examples/bam_read_header_async.rs
@@ -8,27 +8,22 @@ use std::env;
 
 use noodles_bam as bam;
 use noodles_sam as sam;
-use tokio::{fs::File, io};
+use tokio::{
+    fs::File,
+    io::{self, AsyncWriteExt},
+};
 
 #[tokio::main]
 async fn main() -> io::Result<()> {
     let src = env::args().nth(1).expect("missing src");
 
     let mut reader = File::open(src).await.map(bam::AsyncReader::new)?;
-    let raw_header = reader.read_header().await?;
+    let header = reader.read_header().await?;
 
-    if raw_header.is_empty() {
-        let reference_sequences = reader.read_reference_sequences().await?;
+    let mut writer = sam::AsyncWriter::new(io::stdout());
+    writer.write_header(&header).await?;
 
-        let header = sam::Header::builder()
-            .set_reference_sequences(reference_sequences)
-            .build();
-
-        let mut writer = sam::AsyncWriter::new(io::stdout());
-        writer.write_header(&header).await?;
-    } else {
-        print!("{raw_header}");
-    }
+    writer.get_mut().shutdown().await?;
 
     Ok(())
 }