+++ noodles-bam/examples/bam_view_async.rs
@@ -7,15 +7,17 @@ use std::env;
 use futures::TryStreamExt;
 use noodles_bam as bam;
 use noodles_sam as sam;
-use tokio::{fs::File, io};
+use tokio::{
+    fs::File,
+    io::{self, AsyncWriteExt},
+};
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let src = env::args().nth(1).expect("missing src");
 
     let mut reader = File::open(src).await.map(bam::AsyncReader::new)?;
-    let header = reader.read_header().await?.parse()?;
-    reader.read_reference_sequences().await?;
+    let header = reader.read_header().await?;
 
     let mut records = reader.record_bufs(&header);
 
@@ -24,6 +26,8 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     while let Some(record) = records.try_next().await? {
         writer.write_alignment_record(&header, &record).await?;
     }
+
+    writer.get_mut().shutdown().await?;
 
     Ok(())
 }