+++ noodles-bam/src/async/io/reader/header.rs
@@ -1,33 +1,107 @@
 use std::num::NonZeroUsize;
 
 use bstr::BString;
-use noodles_sam::header::{
-    record::value::{map::ReferenceSequence, Map},
-    ReferenceSequences,
+use noodles_sam::{
+    self as sam,
+    header::{
+        record::value::{map::ReferenceSequence, Map},
+        ReferenceSequences,
+    },
 };
-use tokio::io::{self, AsyncRead, AsyncReadExt};
+use tokio::io::{self, AsyncBufRead, AsyncBufReadExt, AsyncRead, AsyncReadExt, BufReader};
 
-use crate::io::reader::bytes_with_nul_to_string;
+use crate::io::reader::{bytes_with_nul_to_string, header::reference_sequences_eq};
 
-pub(super) async fn read_header<R>(reader: &mut R) -> io::Result<String>
+pub(super) async fn read_header<R>(reader: &mut R) -> io::Result<sam::Header>
 where
     R: AsyncRead + Unpin,
 {
-    let l_text = reader.read_u32_le().await.and_then(|n| {
-        usize::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
-    })?;
+    let mut header = read_header_inner(reader).await?;
+    let reference_sequences = read_reference_sequences(reader).await?;
 
-    let mut text = vec![0; l_text];
-    reader.read_exact(&mut text).await?;
+    if header.reference_sequences().is_empty() {
+        *header.reference_sequences_mut() = reference_sequences;
+    } else if !reference_sequences_eq(header.reference_sequences(), &reference_sequences) {
+        return Err(io::Error::new(
+            io::ErrorKind::InvalidData,
+            "SAM header and binary reference sequence dictionaries mismatch",
+        ));
+    }
 
-    // § 4.2 The BAM format (2021-06-03): "Plain header text in SAM; not necessarily
-    // NUL-terminated".
-    bytes_with_nul_to_string(&text).or_else(|_| {
-        String::from_utf8(text).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
-    })
+    Ok(header)
 }
 
-pub(super) async fn read_reference_sequences<R>(reader: &mut R) -> io::Result<ReferenceSequences>
+async fn read_header_inner<R>(reader: &mut R) -> io::Result<sam::Header>
+where
+    R: AsyncRead + Unpin,
+{
+    let l_text = reader.read_u32_le().await.map(u64::from)?;
+
+    let mut parser = sam::header::Parser::default();
+
+    let mut header_reader = BufReader::new(reader.take(l_text));
+    let mut buf = Vec::new();
+
+    while read_line(&mut header_reader, &mut buf).await? != 0 {
+        parser
+            .parse_partial(&buf)
+            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
+    }
+
+    discard_passing(&mut header_reader).await?;
+
+    Ok(parser.finish())
+}
+
+async fn read_line<R>(reader: &mut R, dst: &mut Vec<u8>) -> io::Result<usize>
+where
+    R: AsyncBufRead + Unpin,
+{
+    const NUL: u8 = 0x00;
+    const LINE_FEED: u8 = b'\n';
+    const CARRIAGE_RETURN: u8 = b'\r';
+
+    dst.clear();
+
+    let src = reader.fill_buf().await?;
+
+    if src.is_empty() || src[0] == NUL {
+        return Ok(0);
+    }
+
+    match reader.read_until(LINE_FEED, dst).await? {
+        0 => Ok(0),
+        n => {
+            if dst.ends_with(&[LINE_FEED]) {
+                dst.pop();
+
+                if dst.ends_with(&[CARRIAGE_RETURN]) {
+                    dst.pop();
+                }
+            }
+
+            Ok(n)
+        }
+    }
+}
+
+async fn discard_passing<R>(reader: &mut R) -> io::Result<()>
+where
+    R: AsyncBufRead + Unpin,
+{
+    loop {
+        let src = reader.fill_buf().await?;
+
+        if src.is_empty() {
+            return Ok(());
+        }
+
+        let len = src.len();
+        reader.consume(len);
+    }
+}
+
+async fn read_reference_sequences<R>(reader: &mut R) -> io::Result<ReferenceSequences>
 where
     R: AsyncRead + Unpin,
 {
@@ -71,18 +145,37 @@ where
 
 #[cfg(test)]
 mod tests {
+    use bytes::BufMut;
+
+    use noodles_sam::header::record::value::{
+        map::{self, header::Version},
+        Map,
+    };
+
     use super::*;
+
+    const SQ0_LN: NonZeroUsize = match NonZeroUsize::new(8) {
+        Some(length) => length,
+        None => unreachable!(),
+    };
 
     #[tokio::test]
     async fn test_read_header() -> io::Result<()> {
-        let expected = "@HD\tVN:1.6\n";
-
-        let data_len = expected.len() as u32;
-        let mut data = data_len.to_le_bytes().to_vec();
-        data.extend(expected.as_bytes());
+        let mut data = Vec::new();
+        data.put_u32_le(27); // l_text
+        data.put_slice(b"@HD\tVN:1.6\n@SQ\tSN:sq0\tLN:8\n"); // text
+        data.put_u32_le(1); // n_ref
+        data.put_u32_le(4); // ref[0].l_name
+        data.put_slice(b"sq0\x00"); // ref[0].name
+        data.put_u32_le(8); // ref[0].l_ref
 
         let mut reader = &data[..];
         let actual = read_header(&mut reader).await?;
+
+        let expected = sam::Header::builder()
+            .set_header(Map::<map::Header>::new(Version::new(1, 6)))
+            .add_reference_sequence("sq0", Map::<map::ReferenceSequence>::new(SQ0_LN))
+            .build();
 
         assert_eq!(actual, expected);
 
@@ -91,11 +184,6 @@ mod tests {
 
     #[tokio::test]
     async fn test_read_reference_sequences() -> io::Result<()> {
-        const SQ0_LN: NonZeroUsize = match NonZeroUsize::new(8) {
-            Some(length) => length,
-            None => unreachable!(),
-        };
-
         let data = [
             0x01, 0x00, 0x00, 0x00, // n_ref = 1
             0x04, 0x00, 0x00, 0x00, // ref[0].l_name = 4